Adds refreshing of update blog modals upon close
Add jquery handler so that if user makes changes in input
fields of update blog modal, those will not be retained and if they
click on it again, it will be repopulated with correct content. Did
this with function in blog events file. Also updated handlebars
template for blog to pre-populate update values with existing ones
from database.